Java/Chatty/Chatty-core/src/main/java/org/benti/common/security/jwt/JwtUtils.java
package org.benti.common.security.jwt;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.JwtBuilder;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;

import javax.crypto.spec.SecretKeySpec;
import javax.xml.bind.DatatypeConverter;
import java.security.Key;
import java.util.Date;

/**
 * @author Alejandro Bentivengo
 * @name JwtUtils
 * @date 1/28/2020
 */
public class JwtUtils {

    private static final String JWT_ISSUER = "Chatty";
    private static final String JWT_KEY = "This is the security key for chatty";

    public static String newToken(String subject) {
        SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm.HS512;
        long nowMillis = System.currentTimeMillis();
        Date now = new Date(nowMillis);
        byte[] apiKeySecretBytes = DatatypeConverter.parseBase64Binary(JWT_KEY);
        Key signingKey = new SecretKeySpec(apiKeySecretBytes, signatureAlgorithm.getJcaName());
        JwtBuilder builder = Jwts.builder()
                .setIssuedAt(now)
                .setSubject(subject)
                .setIssuer(JWT_ISSUER)
                .signWith(signatureAlgorithm, signingKey);

        /* No expiration date will be used for now
        if (ttlMillis > 0) {
            long expMillis = nowMillis + ttlMillis;
            Date exp = new Date(expMillis);
            builder.setExpiration(exp);
        }*/

        return builder.compact();
    }

    public static Claims getClaims(String token) {
        return Jwts.parser()
                .setSigningKey(DatatypeConverter.parseBase64Binary(JWT_KEY))
                .parseClaimsJws(token).getBody();
    }
}
